New settings campagnodon_renewal_force_status_id:
When we renew a membership after 1 year,
Civicrm code can't compute the status for membership starting "in the future".
Indead, with Payzen, we receive monthly/annual due 2 weeks in advance...
So we introduce an optional campagnodon_renewal_force_status_id settings.
If set, it will force the status to the settings value ("Current" by default).